pyproject.toml
[project]
name = "narratelean"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [{ name = "narumi", email = "toucans-cutouts0f@icloud.com" }]
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "typer>=0.20.0",
]

[project.scripts]
narratelean = "narratelean.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest-cov>=6.0.0", "pytest>=8.3.3", "ruff>=0.7.3", "ty>=0.0.1a26"]

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"cli.py" = ["B008"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ty.rules]
unresolved-import = "ignore"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.bumpversion]
current_version = "0.1.0"
tag = true
commit = true
pre_commit_hooks = ["uv lock", "git add uv.lock"]
